🎨 style(draculaFast.scss): refactor CSS selectors and remove redundant styles This commit refactors the CSS selectors to improve readability and removes redundant styles. The `html` selector is now combined with the `body` selector to apply the background color and text color to the entire page. The `more-info-link` selector is no longer repeated and the `background-color` style is removed as it is already defined in the `dracula_background` variable. The styles for `#your-speed-message`, `.succeeded`, `#speed-value`, `#speed-units`, `#speed-progress-indicator-icon`, `.measurement-label`, `.bufferbloat-label`, `#bufferbloat-value`, `#upload-value`, `#latency-value`, `.localized`, `table`, `.logo`, `span`, `.powered-by`, `.measurement-details-label`, `.dropbtn`, and `#privacy-link` are removed as they are redundant or not used.
🎨 style(draculaFast.scss): refactor text and hover styles
This commit refactors the text and hover styles in the draculaFast.scss file. The text styles have been consolidated and the hover styles have been updated to include more elements. The color of the speed-progress-indicator has been changed to $dracula_red when it is in a succeeded state. The measurement-details-label has also been updated to use $dracula_red when it is in a succeeded state.